+++ internal/server/handler/register_handler.go
