+++ internal/server/handler/register_handler_test.go
